componetizing
