+++ .vscode/settings.json
@@ -18,5 +18,6 @@
     "titleBar.inactiveBackground": "#1857a499",
     "titleBar.inactiveForeground": "#e7e7e799"
   },
-  "peacock.remoteColor": "#1857a4"
+  "peacock.remoteColor": "#1857a4",
+  "workbench.preferredDarkColorTheme": "Solarized Dark"
 }
