+++ server.js
@@ -1,7 +1,5 @@
 'use strict';
 
-const axios = require('axios');
-
 //first install dotenv, cors, express with npm i _____
 // const weatherData = require('./data/weather.json');
 
@@ -16,6 +14,8 @@ const app = express();
 
 //Bringing in cors
 const cors = require('cors');
+const getMovies = require('./movies');
+const getWeather = require('./weather');
 // const { request } = require('http');
 
 // anyone can make a request to our server
@@ -36,75 +36,10 @@ app.get('/bananas', (req, res) => {
 });
 
 //make a route to our data
-app.get('/weather', async(req, res, next) => {
-  //send our weather data
- 
-  try {
-    // city is our new search query
-    let city = req.query.searchQuery;
-    // this is our URL to take us to the live weather API
-    let liveWeatherUrl = `https://api.weatherbit.io/v2.0/forecast/daily?city=${city}&key=${process.env.REACT_APP_WEATHER_API_KEY}&days=10`;
-    // response to get the live weather data
-    let response = await axios.get(liveWeatherUrl);
-    console.log(response.data);
-    // function to feed the weather data from the specific days into our Forcast Class
-    let descriptions = response.data.data.map(day=> new Forecast(day));
-    // Send out our descriptions data
-    res.status(200).send(descriptions);
-  }
-  catch (error) {
-    next(error);
-  }
-});
+app.get('/weather', getWeather);
 
-app.get('/movies', async(req, res, next) => {
-  //send our weather data
- 
-  try {
-    // city is our new search query
-    let movie = req.query.searchQuery;
-    // this is our URL to take us to the live weather API
-    let liveMovieUrl = `https://api.themoviedb.org/3/search/movie?api_key=${process.env.REACT_APP_MOVIE_API_KEY}&query=${movie}`
-    console.log("Live Movie URL",liveMovieUrl)
-    // response to get the live weather data
-    let response = await axios.get(liveMovieUrl);
-    console.log(response.data);
-    // function to feed the weather data from the specific days into our Forcast Class
-    let descriptions = response.data.results.map(movie=> new Movies(movie));
-    // Send out our descriptions data
-    res.status(200).send(descriptions);
-  }
-  catch (error) {
-    next(error);
-  }
-});
+app.get('/movies', getMovies);
 
-class Movies{
-  constructor(movie){
-    this.title = movie.title,
-    this.overview = movie.overview,
-    this.average_votes = movie.vote_average,
-    this.poster = movie.poster_path,
-    this.popularity = movie.popularity,
-    this.released_on = movie.release_date
-  };
-}
-
-class Forecast {
-  constructor(city) {
-    //takes in city data and breaks it into date and description to be used on the front end. 
-    // This is the data we send to the front
-    this.date = city.datetime;
-    this.description = city.weather.description;
-  }
-
-  // Irrelevent
-  getItems() {
-    return this.city.data.map(items => {
-      return { valid_date: items.valid_date, description: items.weather.description };
-    });
-  }
-}
 // Error response and log
 app.use((error, request, response, next) => {
   console.log(error);
